+++ ja2lib/LoadingScreen.h
@@ -53,7 +53,7 @@ enum {
 extern UINT8 gubLastLoadingScreenID;
 
 // returns the UINT8 ID for the specified sector.
-UINT8 GetLoadScreenID(INT16 sSectorX, INT16 sSectorY, INT8 bSectorZ);
+UINT8 GetLoadScreenID(u8 sSectorX, u8 sSectorY, INT8 bSectorZ);
 
 // sets up the loadscreen with specified ID, and draws it to the FRAME_BUFFER,
 // and refreshing the screen with it.
