+++ ja2lib/Strategic/Assignments.h
@@ -160,7 +160,7 @@ void BuildSectorsWithSoldiersList(void);
 void InitSectorsWithSoldiersList(void);
 
 // is there a soldier in this sector?..only use after BuildSectorsWithSoldiersList is called
-BOOLEAN IsThereASoldierInThisSector(INT16 sSectorX, INT16 sSectorY, INT8 bSectorZ);
+BOOLEAN IsThereASoldierInThisSector(u8 sSectorX, u8 sSectorY, INT8 bSectorZ);
 
 void CheckIfSoldierUnassigned(struct SOLDIERTYPE *pSoldier);
